+++ R/ESTIMATION_FUNS.R
@@ -1735,9 +1735,7 @@ ols_fit = function(y, X, w, correct_0w = FALSE, collin.tol, nthreads, xwx = NULL
     res
 }
 
-
-
-check_conv = function(y, X, fixef_id_list, slope_flag, slope_vars, weights){
+check_conv = function(y, X, fixef_id_list, slope_flag, slope_vars, weights, full = FALSE, fixef_names = NULL){
     # VERY SLOW!!!!
     # IF THIS FUNCTION LASTS => TO BE PORTED TO C++
 
@@ -1748,29 +1746,45 @@ check_conv = function(y, X, fixef_id_list, slope_flag, slope_vars, weights){
 
     Q = length(slope_flag)
 
-    nobs = length(y)
+    use_y = TRUE
     if(length(X) == 1){
         K = 1
+        nobs = length(y)
     } else {
-        K = NCOL(X) + 1
-    }
+        nobs = NROW(X)
 
-    res = list()
-
-    for(k in 1:K){
-        if(k == 1){
-            x = y
-        } else {
-            x = X[, k - 1]
+        if(length(y) <= 1){
+            use_y = FALSE
         }
 
-        res_tmp = c()
+        K = NCOL(X) + use_y
+    }
+
+    info_max = list()
+    info_full = vector("list", K)
+
+    for(k in 1:K){
+        if(k == 1 && use_y){
+            x = y
+        } else if(use_y){
+            x = X[, k - use_y]
+        } else {
+            x = X[, k]
+        }
+
+        info_max_tmp = c()
+        info_full_tmp = list()
         index_slope = 1
         for(q in 1:Q){
             fixef_id = fixef_id_list[[q]]
 
             if(slope_flag[q] >= 0){
-                res_tmp = c(res_tmp, max(abs(tapply(weights * x, fixef_id, mean))))
+                x_means = tapply(weights * x, fixef_id, mean)
+                info_max_tmp = c(info_max_tmp, max(abs(x_means)))
+
+                if(full){
+                    info_full_tmp[[length(info_full_tmp) + 1]] = x_means
+                }
             }
 
             n_slopes = abs(slope_flag[q])
@@ -1780,17 +1794,37 @@ check_conv = function(y, X, fixef_id_list, slope_flag, slope_vars, weights){
 
                     num = tapply(weights * x * var, fixef_id, sum)
                     denom = tapply(weights * var^2, fixef_id, sum)
-                    res_tmp = c(res_tmp, max(abs(num/denom)))
+
+                    x_means = num/denom
+                    info_max_tmp = c(info_max_tmp, max(abs(x_means)))
+
+                    if(full){
+                        info_full_tmp[[length(info_full_tmp) + 1]] = x_means
+                    }
 
                     index_slope = index_slope + 1
                 }
             }
         }
 
-        res[[k]] = res_tmp
+        if(!is.null(fixef_names)){
+            names(info_full_tmp) = fixef_names
+        }
+
+        info_max[[k]] = info_max_tmp
+
+        if(full){
+            info_full[[k]] = info_full_tmp
+        }
     }
 
-    res = do.call("rbind", res)
+    info_max = do.call("rbind", info_max)
+
+    if(full){
+        res = info_full
+    } else {
+        res = info_max
+    }
 
     res
 }